changed unsigned int to unsigned long long

+++ projects/ft_printf/smaller_part_printf/include/ft_printf.h
@@ -35,7 +35,7 @@ typedef struct	s_format
 
 int	ft_printf(const char *, ...);
 int ft_length_nbr(long long nbr);
-int ft_length_x(unsigned int nbr);
+int ft_length_x(unsigned long long nbr);
 int ft_max(int a, int b);
 int print_c(struct s_format *format, va_list a_list);
 int print_str(struct s_format *format, va_list a_list);
@@ -51,10 +51,10 @@ void prec_point(struct s_format *format);
 void x_digits(struct s_format *format, int *max, int *len, char *buff);
 int di_digits(struct s_format *format, int *max, int *len, char *buff);
 void u_digits(struct s_format *format, int *max, int *len, char *buff);
-void x_detect_nbr(unsigned int *nbr, struct s_format *format, char *buff, int *i);
+void x_detect_nbr(unsigned long long *nbr, struct s_format *format, char *buff, int *i);
 int	str_detect(const char *str, struct s_format *format, char *buff);
 int di_detect_nbr(long *nbr, struct s_format *format, char *buff, int *i);
-void putnbr_base(unsigned int *nbr, struct s_format *format, char *buff, int *i);
+void putnbr_base(unsigned long long *nbr, struct s_format *format, char *buff, int *i);
 void di_putnbr_base(long *nbr, struct s_format *format, char *buff, int *i);
 
 #endif
